add `tlc-result-handler` option and update `tla2tools.jar`
`tlc-result-handler` is a var which the user can pass and which will
be run by the child process using the result of the tlc variable.

`tla2tools.jar` now can return the state in `MCVariable` so we can use
it from the JVM directly without having to parse the values from
strings.